python-backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings
supabase
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.3.2
pinecone-client==3.1.0
mammoth==1.6.0
pdfminer.six==20231228
openpyxl==3.1.2
twilio==8.1.0
websockets==12.0
email-validator==2.1.0.post1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
requests==2.31.0
msal==1.26.0
